+++ kitty/options/to-c.h
@@ -192,15 +192,22 @@ text_composition_strategy(PyObject *val, Options *opts) {
     else if (PyUnicode_CompareWithASCIIString(val, "legacy") == 0) {
         opts->text_old_gamma = true;
     } else {
-        DECREF_AFTER_FUNCTION PyObject *parts = PyUnicode_Split(val, NULL, 1);
-        if (PyList_GET_SIZE(parts) != 2) { PyErr_SetString(PyExc_ValueError, "text_rendering_strategy must be of the form number:number"); return; }
-        DECREF_AFTER_FUNCTION PyObject *ga = PyFloat_FromString(PyList_GET_ITEM(parts, 0));
-        if (PyErr_Occurred()) return;
-        opts->text_gamma_adjustment = MAX(0.01f, PyFloat_AsFloat(ga));
-        DECREF_AFTER_FUNCTION PyObject *contrast = PyFloat_FromString(PyList_GET_ITEM(parts, 1));
-        if (PyErr_Occurred()) return;
-        opts->text_contrast = MAX(0.0f, PyFloat_AsFloat(contrast));
-        opts->text_contrast = MIN(100.0f, opts->text_contrast);
+        DECREF_AFTER_FUNCTION PyObject *parts = PyUnicode_Split(val, NULL, 2);
+        int size = PyList_GET_SIZE(parts);
+        if (size < 1 || 2 < size) { PyErr_SetString(PyExc_ValueError, "text_rendering_strategy must be of the form number:[number]"); return; }
+
+        if (size > 0) {
+            DECREF_AFTER_FUNCTION PyObject *ga = PyFloat_FromString(PyList_GET_ITEM(parts, 0));
+            if (PyErr_Occurred()) return;
+            opts->text_gamma_adjustment = MAX(0.01f, PyFloat_AsFloat(ga));
+        }
+
+        if (size > 1) {
+            DECREF_AFTER_FUNCTION PyObject *contrast = PyFloat_FromString(PyList_GET_ITEM(parts, 1));
+            if (PyErr_Occurred()) return;
+            opts->text_contrast = MAX(0.0f, PyFloat_AsFloat(contrast));
+            opts->text_contrast = MIN(100.0f, opts->text_contrast);
+        }
     }
 }